pyproject.toml
[project]
name = "etl-project1"
version = "0.1.0"
description = ""
authors = [
    {name = "Woramateclear",email = "woramate.pong@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "google-cloud-bigquery (>=3.34.0,<4.0.0)",
    "apache-airflow (>=3.0.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
